pyproject.toml
[project]
name = "habit-tracker"
version = "0.1.0"
description = ""
authors = [
    {name = "MasoNord",email = "trotf233@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.133.0,<0.134.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "pydantic[dotenv] (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "redis (>=7.2.0,<8.0.0)"
]

[tool.poetry]
packages = [{include = "habit_tracker", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
